+++ sdk/go/common/resource/plugin/config_source.go
@@ -15,7 +15,7 @@
 package plugin
 
 import (
-	"github.com/pulumi/pulumi/sdk/v2/go/common/resource/config"
+	"github.com/pulumi/pulumi/sdk/v2/go/common/resource"
 	"github.com/pulumi/pulumi/sdk/v2/go/common/tokens"
 )
 
@@ -23,5 +23,5 @@ import (
 // package.
 type ConfigSource interface {
 	// GetPackageConfig returns the set of configuration parameters for the indicated package, if any.
-	GetPackageConfig(pkg tokens.Package) (map[config.Key]string, error)
+	GetPackageConfig(pkg tokens.Package) (resource.PropertyMap, error)
 }